// src/app/global/models/profile-user.ts
export class ProfileUser {
  id: number;
  role_id: number;
  stakeholder_id: number;
  warehouse_id: number;
  city_id: number;
  name: string;
  last_name: string;
  email: string;
  status_id: number;
  document: string;
  phone: string;
  created_at: string;
  updated_at: string;
  business_name: string;
  business: string;
  phone_contact: string;
  send_city_id: number;
  send_city_name: string;
  send_departament_id: number;
  send_departament_name: string;
  address_send: string;
  invoice_city_id: number;
  address_invoice: string;
  invoise_city_name: string;
  invoise_departament_id: number;
  invoise_departament_name: string;


}